Update server.js
Fixed race conditon when deleting webhooks.
Also when the ES integration was @ed to posts it was calling detectIntent the entire text including the @, so detectIntent was receiving for example "@CXBot hello" instead of just "Hello". I have implemented the same solution that I did for the twitter CX integration which was to replace the @ + screenName with an empty string.